Location map (#334)
* Marker is shown

working version of showing a marker

* Multiple markers and centered map

Adds markers to each location of the challenge and provides an info box
when clicking on a marker.
Also centers the map so that all markers are visible.

* Show registered location

When registered to a challenge on-site, show the location in the map.
By default, the registered location is shown. Can be toggled to view all challenge locations.

* Cleanup

Delete google-location-viewer.html. (now replaced by google-multiple-location-viewer)