Challenge: additional resource file
The additional resource file - see coduno/platform#200 - is now shown as part of the
challenge instructions in form of a link. Also, the challenge instructions now can be
viewed while in the code-editor in form of an overlay.
The 'downloadFab' is now always rendered and is used as a trigger to show the
challenge instructions.
In addition, the iron-ajax response is now handled as a string (canonicalName) and
redirects to the correct challenge template when the upload of a new challenge
is finished.
If an upload failed, the exception message is shown as a toast.